auth-backend/pyproject.toml
[project]
name = "authbackend"
version = "0.1.0"
description = "Backend de autentificacion, gestion de usuarios y etc"
authors = [
    {name = "Thehunter101",email = "thehunter121@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.9,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pip (>=25.2,<26.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "opencv-contrib-python (>=4.12.0.88,<5.0.0.0)",
    "insightface (>=0.7.3,<0.8.0)",
    "onnxruntime (>=1.23.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "django-stubs (>=5.2.7,<6.0.0)"
]
